Sync collection registration + core API catch-up (lite→Origin round-trip)

+++ nodedb-lite/src/engine/strict/engine.rs
@@ -65,7 +65,7 @@ impl CollectionState {
         let decoder = TupleDecoder::new(&schema);
         // Initial version maps to current column count.
         let mut version_column_counts = HashMap::new();
-        version_column_counts.insert(schema.version, schema.columns.len());
+        version_column_counts.insert(schema.version as u16, schema.columns.len());
         Self {
             schema,
             encoder,
@@ -96,7 +96,7 @@ impl CollectionState {
 
 /// Encode a PK value into sortable bytes for the storage key.
 ///
-/// Uses big-endian for integers (preserves sort order in redb scans),
+/// Uses big-endian for integers (preserves sort order in B+ tree scans),
 /// raw UTF-8 for strings, and raw bytes for UUIDs.
 fn encode_pk_component(key: &mut Vec<u8>, value: &Value) {
     match value {
